@@ -4,7 +4,7 @@ class GetSwitch extends StatefulWidget {
   final Map<String, dynamic> json;
   final int level;
 
-  const GetSwitch({Key key, @required this.json, @required this.level}) : super(key: key);
+  const GetSwitch({Key? key, required this.json, required this.level}) : super(key: key);
 
   @override
   _GetSwitchState createState() => _GetSwitchState();
@@ -22,19 +22,19 @@ class _GetSwitchState extends State<GetSwitch> {
   @override
   Widget build(BuildContext context) {
 
-    Map<String, dynamic> _json = widget.json;
+    Map<String, dynamic>? _json = widget.json;
 
     if (checked) {
-      _json = (widget.json['variants']['children'] as List)?.firstWhere((element) => (element['name'] as String).contains('checked=true'), orElse: () => null);
+      _json = (widget.json['variants']['children'] as List?)?.firstWhere((element) => (element['name'] as String).contains('checked=true'), orElse: () => null);
     } else {
-      _json = (widget.json['variants']['children'] as List)?.firstWhere((element) => (element['name'] as String).contains('checked=false'), orElse: () => null);
+      _json = (widget.json['variants']['children'] as List?)?.firstWhere((element) => (element['name'] as String).contains('checked=false'), orElse: () => null);
     }
 
     var variable = getVariable(widget.json, 'title');
 
     return GestureDetector(
       onTap: () => setState(() => checked = !checked),
-      child: wrapInstance(_json, getChildrenByLayoutMode(_json, widget.level, variables: [variable]), widget.level + 1),
+      child: wrapInstance(_json!, getChildrenByLayoutMode(_json, widget.level, variables: [variable]), widget.level + 1),
     );
   }
 }
