@@ -1,4 +1,4 @@
-final json_text = '''
+final jsonText = '''
 
 {"type":"FRAME","name":"Frame 86","fills":[{"type":"SOLID","visible":false,"opacity":1,"blendMode":"NORMAL","color":{"r":1,"g":1,"b":1}}],"id":"141:240","parent":{"id":"141:209"},"removed":false,"visible":true,"locked":false,"opacity":1,"blendMode":"PASS_THROUGH","isMask":false,"effects":[],"effectStyleId":"","relativeTransform":[[1,0,0],[0,1,197.5]],"absoluteTransform":[[1,0,1726],[0,1,1951.5]],"x":0,"y":197.5,"width":477,"height":88,"rotation":0,"layoutAlign":"STRETCH","constrainProportions":false,"layoutGrow":0,"children":[{"type":"FRAME","name":"logo","fills":[{"type":"SOLID","visible":false,"opacity":1,"blendMode":"NORMAL","color":{"r":1,"g":1,"b":1}}],"id":"141:241","parent":{"id":"141:240"},"removed":false,"visible":true,"locked":false,"opacity":1,"blendMode":"PASS_THROUGH","isMask":false,"effects":[],"effectStyleId":"","relativeTransform":[[1,0,178.5],[0,1,0]],"absoluteTransform":[[1,0,1904.5],[0,1,1951.5]],"x":178.5,"y":0,"width":119.99999237060547,"height":88,"rotation":0,"layoutAlign":"STRETCH","constrainProportions":false,"layoutGrow":0,"children":[{"type":"VECTOR","name":"Subtract","fills":[{"type":"SOLID","visible":true,"opacity":1,"blendMode":"NORMAL","color":{"r":0.0784313753247261,"g":0.11764705926179886,"b":0.18039216101169586}}],"id":"141:242","svg":"<svg width=\"121\" height=\"89\" viewBox=\"0 0 121 89\" fill=\"none\" xmlns=\"http://www.w3.org/2000/svg\">\n<path d=\"M37.4186 20.8652V18.4442L33.2093 19.647L37.4186 20.8652Z\" fill=\"#141E2E\"/>\n<path d=\"M37.4186 46.2845V43.8636L33.2093 45.0663L37.4186 46.2845Z\" fill=\"#141E2E\"/>\n<path d=\"M34.3267 54.6898C35.2144 54.6898 35.6583 55.2192 35.6583 56.2781V57.2958H33.0256V56.2626C33.0256 55.7281 33.1225 55.3323 33.3164 55.0753C33.5103 54.8183 33.8266 54.6898 34.2654 54.6898H34.3267Z\" fill=\"#141E2E\"/>\n<path d=\"M37.4186 68.7975V66.3765L33.2093 67.5793L37.4186 68.7975Z\" fill=\"#141E2E\"/>\n<path fill-rule=\"evenodd\" clip-rule=\"evenodd\" d=\"M0.5 5.67647C0.5 2.81758 2.82149 0.5 5.68518 0.5H115.315C118.179 0.5 120.5 2.81758 120.5 5.67647V83.3235C120.5 86.1824 118.179 88.5 115.315 88.5H5.68519C2.82149 88.5 0.5 86.1824 0.5 83.3235V5.67647ZM5.68518 3.45798C4.45789 3.45798 3.46296 4.45123 3.46296 5.67647V83.3235C3.46296 84.5488 4.45789 85.542 5.68519 85.542H20.6852V3.45798H5.68518ZM69.6639 85.542H52.7222V56.8741L69.6639 85.542ZM115.315 85.542H105.795L77.537 41.1723L104.495 3.45798H115.315C116.542 3.45798 117.537 4.45124 117.537 5.67647V83.3235C117.537 84.5488 116.542 85.542 115.315 85.542ZM52.7222 3.45798H71.8483L52.7222 31.6934V3.45798ZM30.6684 14.76L36.9288 11.1671H41.6126V7.99051L36.9288 7.99051L30.6684 4.56721V7.58958L34.3573 9.45543L30.6684 11.3213V14.76ZM30.6684 21.482V17.3957L41.6126 13.8953V17.2415L39.5768 17.8274V21.482L41.6126 22.068V25.0133L30.6684 21.482ZM30.6684 36.2774V39.4694L41.6126 36.9096V33.8256L34.6634 32.1139L41.6126 30.4177V27.1178L30.6684 24.5426V27.4724L37.5564 28.9682L30.6684 30.5719V33.3784L37.4951 34.8587L30.6684 36.2774ZM30.6684 46.9014V42.815L41.6126 39.3146V42.6608L39.5768 43.2467V46.9014L41.6126 47.4873V50.4326L30.6684 46.9014ZM30.6684 56.1239V60.4415H41.6126L41.6126 57.2958H37.6941V56.5556L41.6126 54.3197V50.9889L37.0819 53.6258C36.5716 52.2996 35.6124 51.6366 34.2042 51.6366H34.143C32.9899 51.6366 32.1225 52.0221 31.5408 52.7931C30.9592 53.5538 30.6684 54.6641 30.6684 56.1239ZM30.6684 69.4143V65.328L41.6126 61.8276V65.1738L39.5768 65.7597V69.4143L41.6126 70.0003V72.9456L30.6684 69.4143ZM41.7963 78.4961C41.7963 80.398 41.2861 81.8372 40.2656 82.8138C39.3064 83.7699 37.9543 84.2479 36.2094 84.2479H36.0869C34.4236 84.2479 33.0715 83.7133 32.0307 82.6442C31.0102 81.5956 30.5 80.254 30.5 78.6195C30.5 77.17 30.8367 75.9878 31.5102 75.0728C32.2348 74.1476 33.3011 73.6233 34.7093 73.5V76.7536C33.5562 76.877 32.9797 77.4938 32.9797 78.6041C32.9797 79.3237 33.2501 79.8891 33.7909 80.3003C34.3216 80.7115 35.0818 80.9171 36.0716 80.9171H36.1941C37.1839 80.9171 37.9492 80.7166 38.4901 80.3157C39.0207 79.9045 39.286 79.3185 39.286 78.5578C39.286 77.355 38.6891 76.6714 37.4951 76.5069V73.3612C38.8625 73.464 39.934 73.9883 40.7095 74.9341C41.434 75.849 41.7963 77.0363 41.7963 78.4961Z\" fill=\"#141E2E\"/>\n</svg>\n","parent":{"id":"141:241"},"removed":false,"visible":true,"locked":false,"opacity":1,"blendMode":"PASS_THROUGH","isMask":false,"effects":[],"effectStyleId":"","fillStyleId":"S:5a38d41a9fd942c5ee7ee2871ff0a4804cbaa2bd,5:1","strokes":[],"strokeStyleId":"","strokeWeight":16,"strokeAlign":"CENTER","strokeCap":"NONE","strokeJoin":"MITER","strokeMiterLimit":4,"dashPattern":[],"relativeTransform":[[1,0,0],[0,1,0]],"absoluteTransform":[[1,0,1904.5],[0,1,1951.5]],"x":0,"y":0,"width":119.99999237060547,"height":88,"rotation":0,"layoutAlign":"INHERIT","constrainProportions":false,"layoutGrow":0,"exportSettings":[],"constraints":{"horizontal":"SCALE","vertical":"SCALE"},"cornerRadius":0,"cornerSmoothing":0,"vectorNetwork":{"regions":[{"windingRule":"NONZERO","loops":[[44,45,46]]},{"windingRule":"NONZERO","loops":[[68,69,70]]},{"windingRule":"NONZERO","loops":[[83,84,85,86,87,88,89]]},{"windingRule":"NONZERO","loops":[[98,99,100]]},{"windingRule":"EVENODD","loops":[[0,1,2,3,4,5,6,7],[8,9,10,11,12,13],[14,15,16],[17,18,19,20,21,22,23],[24,25,26],[27,28,29,30,31,32,33,34,35],[36,37,38,39,40,41,42,43],[47,48,49,50,51,52,53,54,55,56,57,58,59],[60,61,62,63,64,65,66,67],[71,72,73,74,75,76,77,78,79,80,81,82],[90,91,92,93,94,95,96,97],[101,102,103,104,105,106,107,108,109,110,111,112,113,114,115,116,117,118]]}],"segments":[{"start":0,"end":1,"tangentStart":{"x":-2.8636980056762695,"y":0},"tangentEnd":{"x":0,"y":-2.858886241912842}},{"start":2,"end":0,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":3,"end":2,"tangentStart":{"x":0,"y":-2.858887195587158},"tangentEnd":{"x":2.8636932373046875,"y":0}},{"start":4,"end":3,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":5,"end":4,"tangentStart":{"x":2.8637008666992188,"y":0},"tangentEnd":{"x":0,"y":2.85888671875}},{"start":6,"end":5,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":7,"end":6,"tangentStart":{"x":0,"y":2.85888671875},"tangentEnd":{"x":-2.863697052001953,"y":0}},{"start":1,"end":7,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":8,"end":9,"tangentStart":{"x":0,"y":-1.225236177444458},"tangentEnd":{"x":-1.227299690246582,"y":0}},{"start":10,"end":8,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":11,"end":10,"tangentStart":{"x":-1.2273008823394775,"y":0},"tangentEnd":{"x":0,"y":1.2252349853515625}},{"start":12,"end":11,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":13,"end":12,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":9,"end":13,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":14,"end":15,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":16,"end":14,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":15,"end":16,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":17,"end":18,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":19,"end":17,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":20,"end":19,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":21,"end":20,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":22,"end":21,"tangentStart":{"x":0,"y":-1.2252349853515625},"tangentEnd":{"x":1.227294921875,"y":0}},{"start":23,"end":22,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":18,"end":23,"tangentStart":{"x":1.2273025512695312,"y":0},"tangentEnd":{"x":0,"y":1.2252349853515625}},{"start":24,"end":25,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":26,"end":24,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":25,"end":26,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":27,"end":28,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":29,"end":27,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":30,"end":29,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":31,"end":30,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":32,"end":31,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":33,"end":32,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":34,"end":33,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":35,"end":34,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":28,"end":35,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":36,"end":37,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":38,"end":36,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":39,"end":38,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":40,"end":39,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":41,"end":40,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":42,"end":41,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":43,"end":42,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":37,"end":43,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":44,"end":45,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":46,"end":44,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":45,"end":46,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":47,"end":48,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":49,"end":47,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":50,"end":49,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":51,"end":50,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":52,"end":51,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":53,"end":52,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":54,"end":53,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":55,"end":54,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":56,"end":55,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":57,"end":56,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":58,"end":57,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":59,"end":58,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":48,"end":59,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":60,"end":61,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":62,"end":60,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":63,"end":62,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":64,"end":63,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":65,"end":64,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":66,"end":65,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":67,"end":66,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":61,"end":67,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":68,"end":69,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":70,"end":68,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":69,"end":70,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":71,"end":72,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":73,"end":71,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":74,"end":73,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":75,"end":74,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":76,"end":75,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":77,"end":76,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":78,"end":77,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":79,"end":78,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":80,"end":79,"tangentStart":{"x":1.4082107543945312,"y":0},"tangentEnd":{"x":-0.5102195739746094,"y":-1.3261451721191406}},{"start":81,"end":80,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":82,"end":81,"tangentStart":{"x":0.5816516876220703,"y":-0.7710151672363281},"tangentEnd":{"x":-1.1530990600585938,"y":0}},{"start":72,"end":82,"tangentStart":{"x":0,"y":-1.4597854614257812},"tangentEnd":{"x":-0.5816516876220703,"y":0.7607307434082031}},{"start":83,"end":84,"tangentStart":{"x":0,"y":-1.0588569641113281},"tangentEnd":{"x":0.8877830505371094,"y":0}},{"start":85,"end":83,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":86,"end":85,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":87,"end":86,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":88,"end":87,"tangentStart":{"x":-0.19388580322265625,"y":0.2570075988769531},"tangentEnd":{"x":0,"y":-0.5345687866210938}},{"start":89,"end":88,"tangentStart":{"x":-0.43878936767578125,"y":0},"tangentEnd":{"x":0.19388198852539062,"y":-0.2570037841796875}},{"start":84,"end":89,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":90,"end":91,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":92,"end":90,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":93,"end":92,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":94,"end":93,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":95,"end":94,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":96,"end":95,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":97,"end":96,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":91,"end":97,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":98,"end":99,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":100,"end":98,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":99,"end":100,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":101,"end":102,"tangentStart":{"x":1.0204429626464844,"y":-0.9766159057617188},"tangentEnd":{"x":0,"y":1.9018325805664062}},{"start":103,"end":101,"tangentStart":{"x":1.7449569702148438,"y":0},"tangentEnd":{"x":-0.9592170715332031,"y":0.9560546875}},{"start":104,"end":103,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":105,"end":104,"tangentStart":{"x":1.0408515930175781,"y":1.0691375732421875},"tangentEnd":{"x":-1.6633186340332031,"y":0}},{"start":106,"end":105,"tangentStart":{"x":0,"y":1.634552001953125},"tangentEnd":{"x":-1.0204410552978516,"y":-1.0485763549804688}},{"start":107,"end":106,"tangentStart":{"x":-0.6734905242919922,"y":0.9149322509765625},"tangentEnd":{"x":0,"y":-1.4495010375976562}},{"start":108,"end":107,"tangentStart":{"x":-1.4082107543945312,"y":0.12335968017578125},"tangentEnd":{"x":0.7245140075683594,"y":-0.9252166748046875}},{"start":109,"end":108,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":110,"end":109,"tangentStart":{"x":0,"y":-1.110260009765625},"tangentEnd":{"x":-1.1530990600585938,"y":0.1233673095703125}},{"start":111,"end":110,"tangentStart":{"x":-0.5408363342285156,"y":-0.4112091064453125},"tangentEnd":{"x":0,"y":0.7196121215820312}},{"start":112,"end":111,"tangentStart":{"x":-0.9898300170898438,"y":0},"tangentEnd":{"x":0.5306282043457031,"y":0.4112091064453125}},{"start":113,"end":112,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":114,"end":113,"tangentStart":{"x":-0.5408363342285156,"y":0.4009246826171875},"tangentEnd":{"x":0.9898300170898438,"y":0}},{"start":115,"end":114,"tangentStart":{"x":0,"y":0.7607269287109375},"tangentEnd":{"x":0.5306282043457031,"y":-0.4112091064453125}},{"start":116,"end":115,"tangentStart":{"x":1.1939201354980469,"y":0.16448211669921875},"tangentEnd":{"x":0,"y":-1.2027816772460938}},{"start":117,"end":116,"tangentStart":{"x":0,"y":0},"tangentEnd":{"x":0,"y":0}},{"start":118,"end":117,"tangentStart":{"x":-0.7755393981933594,"y":-0.9457778930664062},"tangentEnd":{"x":1.3673934936523438,"y":0.10280609130859375}},{"start":102,"end":118,"tangentStart":{"x":0,"y":-1.4597854614257812},"tangentEnd":{"x":0.7245140075683594,"y":0.9149322509765625}}],"vertices":[{"x":5.185184955596924,"y":0,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":0,"y":5.176469802856445,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":114.81481170654297,"y":0,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":119.99999237060547,"y":5.176470756530762,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":119.99999237060547,"y":82.82353210449219,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":114.81481170654297,"y":88,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":5.185188293457031,"y":88,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":0,"y":82.82353210449219,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":2.9629628658294678,"y":5.176469802856445,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":5.185184955596924,"y":2.9579832553863525,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":2.9629628658294678,"y":82.82353210449219,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":5.185188293457031,"y":85.0420150756836,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":20.185184478759766,"y":85.0420150756836,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":20.185184478759766,"y":2.9579832553863525,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":52.22222137451172,"y":85.0420150756836,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":69.16393280029297,"y":85.0420150756836,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":52.22222137451172,"y":56.37407684326172,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":105.29547119140625,"y":85.0420150756836,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":114.81481170654297,"y":85.0420150756836,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":77.03703308105469,"y":40.67226791381836,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":103.99539947509766,"y":2.9579832553863525,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":114.81481170654297,"y":2.9579832553863525,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":117.03703308105469,"y":5.176470756530762,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":117.03703308105469,"y":82.82353210449219,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":71.3482666015625,"y":2.9579832553863525,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":52.22222137451172,"y":2.9579832553863525,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":52.22222137451172,"y":31.193374633789062,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":36.42878723144531,"y":10.667084693908691,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":30.16836929321289,"y":14.259997367858887,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11261749267578,"y":10.667084693908691,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11261749267578,"y":7.490511417388916,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":36.42878723144531,"y":7.490511894226074,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":30.16836929321289,"y":4.0672125816345215,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":30.16836929321289,"y":7.089581489562988,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":33.857269287109375,"y":8.955432891845703,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":30.16836929321289,"y":10.821284294128418,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":30.16836929321289,"y":16.895658493041992,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":30.16836929321289,"y":20.98202896118164,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11261749267578,"y":13.395261764526367,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11261749267578,"y":16.741458892822266,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":39.07683563232422,"y":17.327421188354492,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":39.07683563232422,"y":20.98202896118164,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11261749267578,"y":21.567996978759766,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11261749267578,"y":24.513267517089844,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":36.91859817504883,"y":17.94423484802246,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":36.91859817504883,"y":20.365215301513672,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":32.70927047729492,"y":19.147016525268555,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":30.168371200561523,"y":38.96939468383789,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":30.168371200561523,"y":35.777400970458984,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11261749267578,"y":36.40963363647461,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11261749267578,"y":33.3255729675293,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":34.16340255737305,"y":31.613933563232422,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11261749267578,"y":29.917699813842773,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11261749267578,"y":26.617767333984375,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":30.16836929321289,"y":24.042587280273438,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":30.16836929321289,"y":26.972434997558594,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":37.056358337402344,"y":28.468198776245117,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":30.168371200561523,"y":30.0719051361084,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":30.168371200561523,"y":32.87839126586914,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":36.99513244628906,"y":34.35873794555664,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":30.168371200561523,"y":42.31498718261719,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":30.168371200561523,"y":46.40135192871094,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11261749267578,"y":38.81459045410156,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11261749267578,"y":42.16078567504883,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":39.07683563232422,"y":42.74674987792969,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":39.07683563232422,"y":46.40135192871094,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11261749267578,"y":46.98731994628906,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11261749267578,"y":49.93259048461914,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":36.91859817504883,"y":43.363563537597656,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":36.91859817504883,"y":45.784542083740234,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":32.70927429199219,"y":44.566341400146484,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":30.168371200561523,"y":59.94154357910156,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":30.168371200561523,"y":55.623870849609375,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11262130737305,"y":59.94154357910156,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11261749267578,"y":56.79581069946289,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":37.19411849975586,"y":56.79581069946289,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":37.19411849975586,"y":56.05563735961914,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11261749267578,"y":53.819698333740234,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11261749267578,"y":50.48892593383789,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":36.581851959228516,"y":53.125789642333984,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":33.70420455932617,"y":51.13657760620117,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":33.64297866821289,"y":51.13657760620117,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":31.040849685668945,"y":52.29309844970703,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":35.1583366394043,"y":55.77806854248047,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":33.82666015625,"y":54.18978500366211,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":35.1583366394043,"y":56.79581069946289,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":32.52559280395508,"y":56.79581069946289,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":32.52559280395508,"y":55.76264953613281,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":32.81642150878906,"y":54.57529067993164,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":33.76543045043945,"y":54.18978500366211,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":30.168371200561523,"y":64.82795715332031,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":30.168371200561523,"y":68.91432189941406,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11262130737305,"y":61.32755661010742,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11262130737305,"y":64.67375183105469,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":39.07683563232422,"y":65.25971984863281,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":39.07683563232422,"y":68.91432189941406,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11262130737305,"y":69.50029754638672,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.11262130737305,"y":72.44556427001953,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":36.918601989746094,"y":65.87653350830078,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":36.918601989746094,"y":68.29751586914062,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":32.70927429199219,"y":67.07931518554688,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":39.7656364440918,"y":82.31380462646484,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":41.29629898071289,"y":77.99613189697266,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":35.70937728881836,"y":83.74788665771484,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":35.58692169189453,"y":83.74788665771484,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":31.530662536621094,"y":82.14418029785156,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":30,"y":78.11949157714844,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":31.010236740112305,"y":74.57283782958984,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":34.20932388305664,"y":72.99996948242188,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":34.20932388305664,"y":76.2536392211914,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":32.47967529296875,"y":78.10407257080078,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":33.29092788696289,"y":79.80030059814453,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":35.571617126464844,"y":80.4171142578125,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":35.694068908691406,"y":80.4171142578125,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":37.99006652832031,"y":79.81571960449219,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":38.7860107421875,"y":78.05781555175781,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":36.99513244628906,"y":76.00691986083984,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":36.99513244628906,"y":72.8611831665039,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"},{"x":40.209529876708984,"y":74.43405151367188,"strokeCap":"NONE","strokeJoin":"MITER","cornerRadius":0,"handleMirroring":"NONE"}]},"vectorPaths":[{"windingRule":"NONZERO","data":"M 36.91859817504883 20.365215301513672 L 36.91859817504883 17.94423484802246 L 32.70927047729492 19.147016525268555 L 36.91859817504883 20.365215301513672 Z"},{"windingRule":"NONZERO","data":"M 36.91859817504883 45.784542083740234 L 36.91859817504883 43.363563537597656 L 32.70927429199219 44.566341400146484 L 36.91859817504883 45.784542083740234 Z"},{"windingRule":"NONZERO","data":"M 33.82666015625 54.18978500366211 C 34.71444320678711 54.18978500366211 35.1583366394043 54.71921157836914 35.1583366394043 55.77806854248047 L 35.1583366394043 56.79581069946289 L 32.52559280395508 56.79581069946289 L 32.52559280395508 55.76264953613281 C 32.52559280395508 55.22808074951172 32.622535705566406 54.832298278808594 32.81642150878906 54.57529067993164 C 33.01030349731445 54.31828689575195 33.32664108276367 54.18978500366211 33.76543045043945 54.18978500366211 L 33.82666015625 54.18978500366211 Z"},{"windingRule":"NONZERO","data":"M 36.918601989746094 68.29751586914062 L 36.918601989746094 65.87653350830078 L 32.70927429199219 67.07931518554688 L 36.918601989746094 68.29751586914062 Z"},{"windingRule":"EVENODD","data":"M 0 5.176469802856445 C 0 2.3175835609436035 2.3214869499206543 0 5.185184955596924 0 L 114.81481170654297 0 C 117.67850494384766 0 119.99999237060547 2.3175835609436035 119.99999237060547 5.176470756530762 L 119.99999237060547 82.82353210449219 C 119.99999237060547 85.68241882324219 117.67851257324219 88 114.81481170654297 88 L 5.185188293457031 88 C 2.321491241455078 88 0 85.68241882324219 0 82.82353210449219 L 0 5.176469802856445 Z M 5.185184955596924 2.9579832553863525 C 3.957885265350342 2.9579832553863525 2.9629628658294678 3.9512336254119873 2.9629628658294678 5.176469802856445 L 2.9629628658294678 82.82353210449219 C 2.9629628658294678 84.04876708984375 3.9578874111175537 85.0420150756836 5.185188293457031 85.0420150756836 L 20.185184478759766 85.0420150756836 L 20.185184478759766 2.9579832553863525 L 5.185184955596924 2.9579832553863525 Z M 69.16393280029297 85.0420150756836 L 52.22222137451172 85.0420150756836 L 52.22222137451172 56.37407684326172 L 69.16393280029297 85.0420150756836 Z M 114.81481170654297 85.0420150756836 L 105.29547119140625 85.0420150756836 L 77.03703308105469 40.67226791381836 L 103.99539947509766 2.9579832553863525 L 114.81481170654297 2.9579832553863525 C 116.04210662841797 2.9579832553863525 117.03703308105469 3.951235771179199 117.03703308105469 5.176470756530762 L 117.03703308105469 82.82353210449219 C 117.03703308105469 84.04876708984375 116.0421142578125 85.0420150756836 114.81481170654297 85.0420150756836 Z M 52.22222137451172 2.9579832553863525 L 71.3482666015625 2.9579832553863525 L 52.22222137451172 31.193374633789062 L 52.22222137451172 2.9579832553863525 Z M 30.16836929321289 14.259997367858887 L 36.42878723144531 10.667084693908691 L 41.11261749267578 10.667084693908691 L 41.11261749267578 7.490511417388916 L 36.42878723144531 7.490511894226074 L 30.16836929321289 4.0672125816345215 L 30.16836929321289 7.089581489562988 L 33.857269287109375 8.955432891845703 L 30.16836929321289 10.821284294128418 L 30.16836929321289 14.259997367858887 Z M 30.16836929321289 20.98202896118164 L 30.16836929321289 16.895658493041992 L 41.11261749267578 13.395261764526367 L 41.11261749267578 16.741458892822266 L 39.07683563232422 17.327421188354492 L 39.07683563232422 20.98202896118164 L 41.11261749267578 21.567996978759766 L 41.11261749267578 24.513267517089844 L 30.16836929321289 20.98202896118164 Z M 30.168371200561523 35.777400970458984 L 30.168371200561523 38.96939468383789 L 41.11261749267578 36.40963363647461 L 41.11261749267578 33.3255729675293 L 34.16340255737305 31.613933563232422 L 41.11261749267578 29.917699813842773 L 41.11261749267578 26.617767333984375 L 30.16836929321289 24.042587280273438 L 30.16836929321289 26.972434997558594 L 37.056358337402344 28.468198776245117 L 30.168371200561523 30.0719051361084 L 30.168371200561523 32.87839126586914 L 36.99513244628906 34.35873794555664 L 30.168371200561523 35.777400970458984 Z M 30.168371200561523 46.40135192871094 L 30.168371200561523 42.31498718261719 L 41.11261749267578 38.81459045410156 L 41.11261749267578 42.16078567504883 L 39.07683563232422 42.74674987792969 L 39.07683563232422 46.40135192871094 L 41.11261749267578 46.98731994628906 L 41.11261749267578 49.93259048461914 L 30.168371200561523 46.40135192871094 Z M 30.168371200561523 55.623870849609375 L 30.168371200561523 59.94154357910156 L 41.11262130737305 59.94154357910156 L 41.11261749267578 56.79581069946289 L 37.19411849975586 56.79581069946289 L 37.19411849975586 56.05563735961914 L 41.11261749267578 53.819698333740234 L 41.11261749267578 50.48892593383789 L 36.581851959228516 53.125789642333984 C 36.071632385253906 51.799644470214844 35.1124153137207 51.13657760620117 33.70420455932617 51.13657760620117 L 33.64297866821289 51.13657760620117 C 32.4898796081543 51.13657760620117 31.622501373291016 51.5220832824707 31.040849685668945 52.29309844970703 C 30.459197998046875 53.053829193115234 30.168371200561523 54.164085388183594 30.168371200561523 55.623870849609375 Z M 30.168371200561523 68.91432189941406 L 30.168371200561523 64.82795715332031 L 41.11262130737305 61.32755661010742 L 41.11262130737305 64.67375183105469 L 39.07683563232422 65.25971984863281 L 39.07683563232422 68.91432189941406 L 41.11262130737305 69.50029754638672 L 41.11262130737305 72.44556427001953 L 30.168371200561523 68.91432189941406 Z M 41.29629898071289 77.99613189697266 C 41.29629898071289 79.89796447753906 40.78607940673828 81.33718872070312 39.7656364440918 82.31380462646484 C 38.806419372558594 83.26985931396484 37.4543342590332 83.74788665771484 35.70937728881836 83.74788665771484 L 35.58692169189453 83.74788665771484 C 33.92360305786133 83.74788665771484 32.57151412963867 83.21331787109375 31.530662536621094 82.14418029785156 C 30.510221481323242 81.0956039428711 30 79.75404357910156 30 78.11949157714844 C 30 76.66999053955078 30.336746215820312 75.4877700805664 31.010236740112305 74.57283782958984 C 31.734750747680664 73.64762115478516 32.80111312866211 73.12332916259766 34.20932388305664 72.99996948242188 L 34.20932388305664 76.2536392211914 C 33.05622482299805 76.37700653076172 32.47967529296875 76.99381256103516 32.47967529296875 78.10407257080078 C 32.47967529296875 78.82368469238281 32.750091552734375 79.38909149169922 33.29092788696289 79.80030059814453 C 33.821556091308594 80.21150970458984 34.581787109375 80.4171142578125 35.571617126464844 80.4171142578125 L 35.694068908691406 80.4171142578125 C 36.68389892578125 80.4171142578125 37.4492301940918 80.21664428710938 37.99006652832031 79.81571960449219 C 38.520694732666016 79.40451049804688 38.7860107421875 78.81854248046875 38.7860107421875 78.05781555175781 C 38.7860107421875 76.85503387451172 38.18905258178711 76.17140197753906 36.99513244628906 76.00691986083984 L 36.99513244628906 72.8611831665039 C 38.362525939941406 72.9639892578125 39.433990478515625 73.48827362060547 40.209529876708984 74.43405151367188 C 40.934043884277344 75.34898376464844 41.29629898071289 76.53634643554688 41.29629898071289 77.99613189697266 Z"}],"handleMirroring":"NONE","reactions":[]}],"exportSettings":[],"fillStyleId":"","strokes":[],"strokeStyleId":"","strokeWeight":1,"strokeAlign":"INSIDE","strokeCap":"NONE","strokeJoin":"MITER","strokeMiterLimit":4,"dashPattern":[],"cornerRadius":0,"cornerSmoothing":0,"topLeftRadius":0,"topRightRadius":0,"bottomLeftRadius":0,"bottomRightRadius":0,"paddingLeft":0,"paddingRight":0,"paddingTop":0,"paddingBottom":0,"primaryAxisAlignItems":"MIN","counterAxisAlignItems":"MIN","primaryAxisSizingMode":"AUTO","layoutGrids":[],"gridStyleId":"","backgrounds":[{"type":"SOLID","visible":false,"opacity":1,"blendMode":"NORMAL","color":{"r":1,"g":1,"b":1}}],"backgroundStyleId":"","clipsContent":false,"guides":[],"expanded":true,"constraints":{"horizontal":"MIN","vertical":"MIN"},"layoutMode":"NONE","counterAxisSizingMode":"FIXED","horizontalPadding":0,"verticalPadding":0,"itemSpacing":0,"overflowDirection":"NONE","numberOfFixedChildren":0,"overlayPositionType":"CENTER","overlayBackground":{"type":"NONE"},"overlayBackgroundInteraction":"NONE","reactions":[]}],"exportSettings":[],"fillStyleId":"","strokes":[],"strokeStyleId":"","strokeWeight":1,"strokeAlign":"INSIDE","strokeCap":"NONE","strokeJoin":"MITER","strokeMiterLimit":4,"dashPattern":[],"cornerRadius":0,"cornerSmoothing":0,"topLeftRadius":0,"topRightRadius":0,"bottomLeftRadius":0,"bottomRightRadius":0,"paddingLeft":0,"paddingRight":0,"paddingTop":0,"paddingBottom":0,"primaryAxisAlignItems":"CENTER","counterAxisAlignItems":"CENTER","primaryAxisSizingMode":"FIXED","layoutGrids":[],"gridStyleId":"","backgrounds":[{"type":"SOLID","visible":false,"opacity":1,"blendMode":"NORMAL","color":{"r":1,"g":1,"b":1}}],"backgroundStyleId":"","clipsContent":false,"guides":[],"expanded":true,"constraints":{"horizontal":"MIN","vertical":"MIN"},"layoutMode":"HORIZONTAL","counterAxisSizingMode":"AUTO","horizontalPadding":0,"verticalPadding":0,"itemSpacing":10,"overflowDirection":"NONE","numberOfFixedChildren":0,"overlayPositionType":"CENTER","overlayBackground":{"type":"NONE"},"overlayBackgroundInteraction":"NONE","reactions":[]}
 
