@@ -1,6 +1,6 @@
 import 'package:flutter_visible/imports.dart';
 
-Color getColor(Map<String, dynamic> color, {num opacity = 1}) {
+Color getColor(Map<String, dynamic>? color, {num? opacity = 1}) {
   if (viewDebugProps) print('getColor');
 
   if (color == null) return Colors.transparent;
@@ -9,6 +9,6 @@ Color getColor(Map<String, dynamic> color, {num opacity = 1}) {
     (color['r'] * 255).round(),
     (color['g'] * 255).round(),
     (color['b'] * 255).round(),
-    color['a'] ?? opacity,
+    color['a'] ?? opacity as double,
   );
 }
